// ts/enums/EntityType.ts
export enum EntityType {
    PLAYERS = 0,
    FIRE = 1,
    WORKBENCH = 2,
    SEED = 3,
    WALL = 4,
    SPIKE = 5,
    BIG_FIRE = 6,
    STONE_WALL = 7,
    GOLD_WALL = 8,
    DIAMOND_WALL = 9,
    WOOD_DOOR = 10,
    CHEST = 11,
    STONE_SPIKE = 12,
    GOLD_SPIKE = 13,
    DIAMOND_SPIKE = 14,
    STONE_DOOR = 15,
    GOLD_DOOR = 16,
    DIAMOND_DOOR = 17,
    FURNACE = 18,
    AMETHYST_WALL = 19,
    AMETHYST_SPIKE = 20,
    AMETHYST_DOOR = 21,
    RESURRECTION = 22,
    EMERALD_MACHINE = 23,
    EXTRACTOR_MACHINE_STONE = 24,
    EXTRACTOR_MACHINE_GOLD = 25,
    EXTRACTOR_MACHINE_DIAMOND = 26,
    EXTRACTOR_MACHINE_AMETHYST = 27,
    EXTRACTOR_MACHINE_REIDITE = 28,
    TOTEM = 29,
    BRIDGE = 30,
    WHEAT_SEED = 31,
    WINDMILL = 32,
    PLOT = 33,
    BREAD_OVEN = 34,
    WELL = 35,
    SIGN = 36,
    PUMPKIN_SEED = 37,
    ROOF = 38,
    GARLIC_SEED = 39,
    THORNBUSH_SEED = 40,
    BED = 41,
    GARLAND = 42,
    TOMATO_SEED = 43,
    CARROT_SEED = 44,
    WOOD_DOOR_SPIKE = 45,
    STONE_DOOR_SPIKE = 46,
    GOLD_DOOR_SPIKE = 47,
    DIAMOND_DOOR_SPIKE = 48,
    AMETHYST_DOOR_SPIKE = 49,
    REIDITE_WALL = 50,
    REIDITE_DOOR = 51,
    REIDITE_SPIKE = 52,
    REIDITE_DOOR_SPIKE = 53,
    WATERMELON_SEED = 54,
    ALOE_VERA_SEED = 55,
    WOOD_TOWER = 56,
    WOLF = 60,
    SPIDER = 61,
    FOX = 62,
    BEAR = 63,
    DRAGON = 64,
    PIRANHA = 65,
    KRAKEN = 66,
    CRAB = 67,
    FLAME = 68,
    LAVA_DRAGON = 69,
    BOAR = 70,
    CRAB_BOSS = 71,
    BABY_DRAGON = 72,
    BABY_LAVA = 73,
    HAWK = 74,
    VULTURE = 75,
    SAND_WORM = 76,
    BABY_MAMMOTH = 77,
    MAMMOTH = 78,
    WHEAT_MOB = 79,
    RABBIT = 80,
    TREASURE_CHEST = 81,
    DEAD_BOX = 82,
    PUMPKIN_MOB = 83,
    GARLIC_MOB = 84,
    THORNBUSH_MOB = 85,
    CRATE = 86,
    GIFT = 87,
    PENGUIN = 88,
    ALOE_VERA_MOB = 89,
    FIREFLY = 90,
    SPELL = 91,
    FRUIT = 100
}